Integrate the repository layer with the Customer entity and clean up and move CustomerRepositoryImpl file
