+++ src/main/java/com/quintrix/carportal/customer/CustomerApplication.java
@@ -2,8 +2,9 @@ package com.quintrix.carportal.customer;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = DataSourceAutoConfiguration.class)
 public class CustomerApplication {
 
 	public static void main(String[] args) {
